requirements.txt
click==8.1.3
numpy==1.24.3
Pillow-SIMD==9.0.0.post1
tqdm==4.65.0
